dom/indexedDB/test/test_objectStore_remove_values.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">

function continueTest(aEvent)
{
  SimpleTest.executeSoon(function() {
    gGen.send(aEvent);
  });
}

function errorHandler(aEvent)
{
  ok(false, "indexedDB error (" + aEvent.code + "): " + aEvent.message);
  SimpleTest.executeSoon(finishTest);
}

function unexpectedSuccessHandler()
{
  ok(false, "Got success, but did not expect it!");
  SimpleTest.executeSoon(finishTest);
}

function ExpectError(aCode)
{
  this._code = aCode;
}
ExpectError.prototype = {
  handleEvent: function(aEvent)
  {
    is(this._code, aEvent.code, "Expected error was thrown.");
    continueTest(aEvent);
  },
};

function testSteps()
{
  const IDBObjectStore = Components.interfaces.nsIIDBObjectStore;
  const IDBDatabaseException = Components.interfaces.nsIIDBDatabaseException;
  const name = window.location.pathname;
  const description = "My Test Database";

  var request = indexedDB.open(name, description);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  var event = yield;

  var db = event.result;

  // Check removal for inline keys, with a key generator.
  var objectStoreName = "inline key; key generator";
  var request = db.createObjectStore(objectStoreName, "id", true);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  request = db.openObjectStore(objectStoreName);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  var objectStore = event.result;
  var storedObject = {name: "Lincoln"};
  request = objectStore.put(storedObject);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  var id = event.result;
  request = objectStore.get(id);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  // Sanity check!
  SimpleTest.is(storedObject.name, event.result.name,
                "The correct object was stored.");

  request = objectStore.remove(id);
  request.onerror = errorHandler;
  request.onsuccess = continueTest;
  event = yield;

  // Make sure it was removed.
  request = objectStore.get(id);
  request.onerror = new ExpectError(IDBDatabaseException.NOT_FOUND_ERR);
  request.onsuccess = unexpectedSuccessHandler;
  event = yield;

  // TODO test the other cases!

  SimpleTest.executeSoon(finishTest);
  yield;
}

function finishTest()
{
  gGen.close();
  SimpleTest.finish();
}

var gGen = testSteps();
function run_test()
{
  SimpleTest.waitForExplicitFinish();
  continueTest();
}

  </script>

</head>

<body onload="run_test();">
</body>
</html>
